+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.0
+
+- Renames `compute` and `computeBest` to `fit` and `bestFit`.
+- Adds `Clusters.predict` and `Clusters.inertia`.
+- Allows dimensions of the input data to be ignored.
+- Allows specifying a method of selecting initial means.
+- Translates and scales input points before fitting.
+- Adds tests.
+
 ## 0.1.0
 
 - Initial version.